divided up the testGame files
